pyproject.toml
[tool.poetry]
name = "mushishi"
version = "0.2.0"
description = "Discord bot in python"
authors = ["Grayson Miller <grayson.miller124@gmail.com>"]
license = "GPLv3"
readme = "README.md"
packages = [{include = "mushishi", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
nltk = "^3.7"
"discord.py" = "^2.0.1"
numpy = "^1.23.4"
pandas = "^1.5.1"
networkx = "^2.8.7"
matplotlib = "^3.6.1"
uvloop = {version = "^0.17.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pylint = "^2.15.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mushi = "mushishi.__main__:main"

